requirements.txt
jinja2
eth-brownie
pydantic
beartype
tinydb
filelock
simular-evm
simular-eth-stubs
# dill